pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-study-buddy"
version = "0.1.0"
description = "CLI-first PDF RAG pipeline using LangChain, Gemini, and ChromaDB."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "langchain>=0.3.0",
  "langchain-google-genai>=2.0.0",
  "langchain-chroma>=0.2.0",
  "chromadb>=0.5.0",
  "pypdf>=5.0.0",
  "python-dotenv>=1.0.1",
  "streamlit>=1.40.0",
  "typer>=0.12.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
study-buddy = "study_buddy.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
study_buddy = ["prompts/*.txt"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
